+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@ This a informal changelog so i can keep track of what im doing.
     - Disk Scans have some info, but I can't see it because it runs too fast lmao
     - Chapter downloads have per-page download progress shown
 - Mangakakalot provider is now working, since we can bypass cloudflare properly
+- Fix mangadex issues with scanlators and official publisher stuff
 
 ## 2026-03-23